Sentiment Analysis in Finance
We assess the effectiveness of many different preprocessing techniques and machine learning algorithms for sentiment analysis in finance using a sizable dataset of messages posted on StockTwits. We discover that the performance of sentiment categorization is greatly enhanced with the addition of bigrams and emojis. We also present empirical evidence showing that the association between investor sentiment and stock returns is significantly influenced by the size of the dataset and the preprocessing technique. Although there is a strong correlation between stock returns and investor mood, we do not find that investor sentiment as inferred from messages posted on social media is useful in forecasting the return of large capitalization equities on a daily basis.
 main.